Change notes to correct octaves
Because the normal convention is for octaves to start at C (not at A), all notes except A, A♯, and B must be moved up one octave in the naming scheme.

Unfortunately, this will be a breaking change for any library users who referred to those notes by their previous name.